// src/modules/search/enums/sort-field.enum.ts
/**
 * Sort Field Enum
 *
 * Defines the fields that can be used for sorting search results
 */
export enum SortField {
  RELEVANCE = 'relevance',
  NAME = 'name',
  START_DATE = 'startDate',
  END_DATE = 'endDate',
  BUDGET = 'budget',
  PERCENT_COMPLETE = 'percentComplete',
  DISTANCE = 'distance',
}